pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf_project"
version = "0.1.0"
description = "Attendance report PDF variation generator"
readme = "README.md"
authors = [{ name = "Your Name" }]
requires-python = ">=3.8"
dependencies = [
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
    "img2table>=1.2.0",
    "Pillow>=10.0.0",
    "reportlab>=4.0.0",
    "python-bidi>=0.4.2",
    "arabic-reshaper>=3.0.0",
]

[project.scripts]
pdf-vary  = "pdf_project.main:cli"
pdf-learn = "pdf_project.learner_cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pdf_project = ["assets/*.ttf"]

[tool.pytest.ini_options]
testpaths = ["tests"]
